Refactor ctrl selection
- Make selection a signed int to use -1 for unselected
- Only use 1 selection for left+right (Part 1 of #440)
- Use unsigned selection only where we have actually an index and no invalid ones (choose item)

+++ src/controls/ctrlComboBox.h
@@ -46,7 +46,7 @@ class ctrlComboBox : public Window
         bool Msg_RightDown(const MouseCoords& mc) override;
         bool Msg_WheelUp(const MouseCoords& mc) override;
         bool Msg_WheelDown(const MouseCoords& mc) override;
-        void Msg_ListSelectItem(const unsigned int ctrl_id, const unsigned short selection) override;
+        void Msg_ListSelectItem(const unsigned int ctrl_id, const int selection) override;
 
     protected:
         bool Draw_() override;
