+++ src/ingameWindows/iwChat.cpp
@@ -60,7 +60,7 @@ void iwChat::Msg_PaintBefore()
     GetCtrl<ctrlEdit>(0)->SetFocus();
 }
 
-void iwChat::Msg_OptionGroupChange(const unsigned int  /*ctrl_id*/, const unsigned short selection)
+void iwChat::Msg_OptionGroupChange(const unsigned int  /*ctrl_id*/, const int selection)
 {
     chat_dest = static_cast<unsigned char>(selection);
     GetCtrl<ctrlEdit>(0)->SetFocus();
