+++ src/ingameWindows/iwDirectIPConnect.cpp
@@ -186,7 +186,7 @@ void iwDirectIPConnect::Msg_ButtonClick(const unsigned int ctrl_id)
  *
  *  @author FloSoft
  */
-void iwDirectIPConnect::Msg_OptionGroupChange(const unsigned int ctrl_id, const unsigned short selection)
+void iwDirectIPConnect::Msg_OptionGroupChange(const unsigned int ctrl_id, const int selection)
 {
     switch(ctrl_id)
     {
