+++ src/ingameWindows/iwDirectIPCreate.cpp
@@ -131,7 +131,7 @@ void iwDirectIPCreate::Msg_EditEnter(const unsigned int ctrl_id)
  *
  *  @author FloSoft
  */
-void iwDirectIPCreate::Msg_OptionGroupChange(const unsigned int ctrl_id, const unsigned short selection)
+void iwDirectIPCreate::Msg_OptionGroupChange(const unsigned int ctrl_id, const int selection)
 {
     switch(ctrl_id)
     {
