+++ src/ingameWindows/iwSave.h
@@ -39,7 +39,7 @@ class iwSaveLoad : public IngameWindow
 
         void Msg_EditEnter(const unsigned int ctrl_id) override;
         void Msg_ButtonClick(const unsigned int ctrl_id) override;
-        void Msg_TableSelectItem(const unsigned int ctrl_id, const unsigned short selection) override;
+        void Msg_TableSelectItem(const unsigned int ctrl_id, const int selection) override;
 
         /// Callbackfunktion zum Eintragen eines Spielstandes in die Tabelle
         static void FillSaveTable(const std::string& filePath, void* param);
@@ -54,7 +54,7 @@ class iwSave: public iwSaveLoad
         // Speichert Datei
         void SaveLoad() override;
 
-        void Msg_ComboSelectItem(const unsigned int ctrl_id, const unsigned short selection) override;
+        void Msg_ComboSelectItem(const unsigned int ctrl_id, const int selection) override;
 };
 
 
@@ -67,7 +67,7 @@ class iwLoad: public iwSaveLoad
 
     private:
         /// Handle double click on the table
-        void Msg_TableChooseItem(const unsigned ctrl_id, const unsigned short selection) override;
+        void Msg_TableChooseItem(const unsigned ctrl_id, const unsigned selection) override;
 
         // Läd Datei
         void SaveLoad() override;
