+++ src/ingameWindows/iwStatistics.cpp
@@ -199,7 +199,7 @@ void iwStatistics::Msg_ButtonClick(const unsigned int ctrl_id)
  *
  *  @author jh
  */
-void iwStatistics::Msg_OptionGroupChange(const unsigned int ctrl_id, const unsigned short selection)
+void iwStatistics::Msg_OptionGroupChange(const unsigned int ctrl_id, const int selection)
 {
     switch(ctrl_id)
     {
