+++ src/ingameWindows/iwStatistics.h
@@ -46,7 +46,7 @@ class iwStatistics : public IngameWindow
 
         void Msg_ButtonClick(const unsigned int ctrl_id) override;
         void Msg_PaintAfter() override;
-        void Msg_OptionGroupChange(const unsigned int ctrl_id, const unsigned short selection) override;
+        void Msg_OptionGroupChange(const unsigned int ctrl_id, const int selection) override;
         void DrawStatistic(StatisticType type);
         void DrawAxis();
 };
