+++ src/ingameWindows/iwTrade.h
@@ -23,7 +23,7 @@ class iwTrade : public IngameWindow
         void Msg_PaintBefore() override;
         void Msg_PaintAfter() override;
         void Msg_ButtonClick(const unsigned int ctrl_id) override;
-        void Msg_ComboSelectItem(const unsigned ctrl_id, const unsigned short selection) override;
+        void Msg_ComboSelectItem(const unsigned ctrl_id, const int selection) override;
         unsigned GetPossibleTradeAmount(const Job job) const;
         unsigned GetPossibleTradeAmount(const GoodType good) const;
 };
